requirements.txt
boto3==1.15.15
imageio==2.8.0
Keras==2.4.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.2.1
networkx==2.4
NNetwork==0.0.1
numpy==1.18.5
pandas==1.0.3
pandas-datareader==0.8.1
Pillow==7.1.2
psutil==5.7.0
scikit-image==0.17.2
scikit-learn==0.23.1
scikit-plot==0.3.7
scipy==1.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
torch==1.5.0
torchvision==0.6.0
torchviz==0.0.1
tqdm==4.46.1
